update uftpd, make server high priority
This fixes uftpd notifying about non errors which probably
caused a race condition on the spi bus.

It appears that when main_task and the ftp_server task
have the same priority they can interfer with each other.
As a result the server couldn't access the filesystem and
I got errors like these:

    E (18884) sdmmc_cmd: sdmmc_read_sectors_dma: sdmmc_send_cmd returned 0x107
    E (18884) diskio_sdmmc: sdmmc_read_blocks failed (263)
    E (27184) sdmmc_cmd: sdmmc_read_sectors_dma: sdmmc_send_cmd returned 0x107
    E (27185) diskio_sdmmc: sdmmc_read_blocks failed (263)

I assume that this might be because the display got updated in
the same moment as the fat driver read from the sd card.
But that shouldn't happen as the spi_master driver from esp-idf
at least promises to be threadsafe.

After removing main/main.c:229 which contained

	    ui_display_ftp_status(event.ftp);

the error disapperad which reinforces my hypothesis.